+++ src/CredentialsServiceProvider.php
@@ -76,7 +76,11 @@ class CredentialsServiceProvider extends ServiceProvider
         $this->registerUserProvider();
         $this->registerGroupProvider();
         $this->registerCredentials();
-        $this->registerCommandSubscriber();
+
+        if ($app['config']['graham-campbell/core::commands']) {
+            $this->registerCommandSubscriber();
+        }
+
         $this->registerAccountController();
         $this->registerLoginController();
         $this->registerRegistrationController();