+++ src/CredentialsServiceProvider.php
@@ -45,11 +45,35 @@ class CredentialsServiceProvider extends ServiceProvider
     {
         $this->package('graham-campbell/credentials', 'graham-campbell/credentials', __DIR__);
 
+        $this->setupBlade();
+
         $this->setupViewer();
 
         include __DIR__.'/routes.php';
         include __DIR__.'/filters.php';
         include __DIR__.'/listeners.php';
+    }
+
+    /**
+     * Setup the blade compiler class.
+     *
+     * @return void
+     */
+    protected function setupBlade()
+    {
+        $blade = $this->app['view']->getEngineResolver()->resolve('blade')->getCompiler();
+
+        $blade->extend(function ($value, $compiler) {
+            $pattern = $compiler->createMatcher('auth');
+            $replace = '$1<?php if (\GrahamCampbell\Credentials\Facades\Credentials::check() && \GrahamCampbell\Credentials\Facades\Credentials::hasAccess$2): ?>';
+            return preg_replace($pattern, $replace, $value);
+        });
+
+        $blade->extend(function ($value, $compiler) {
+            $pattern = $compiler->createPlainMatcher('endauth');
+            $replace = '$1<?php endif; ?>$2';
+            return preg_replace($pattern, $replace, $value);
+        });
     }
 
     /**