+++ CHANGELOG.md
@@ -12,6 +12,7 @@ CHANGE LOG
 * PHP 5.6 and HHVM compatibility
 * Updated to Bootstrap 3.1
 * Controller refactoring
+* Moved home path to core package
 * Command improvements
 * Composer tweaks
 * Added more unit tests